lib/schematools/example/requirements.txt
django-restframework
